refactor: (WIP) add soft deprecation of DataBridgesShape
