+++ data_bridges_knots/client.py
@@ -15,7 +15,7 @@ from data_bridges_client.token import WfpApiToken
 
 from data_bridges_knots.helpers import get_adm0_code
 
-from .endpoints.prices import get_prices, get_exchange_rates
+# from data.endpoints.prices import get_prices, get_exchange_rates
 
 logname = "data_bridges_api_calls.log"
 logging.basicConfig(
@@ -78,8 +78,7 @@ def config_from_env() -> Dict:
 
     return config
 
-
-class DataBridgesShapes:
+class DataBridgesKnots:
     """Interface to the Data Bridges API.
 
     Provides methods for fetching market prices, exchange rates, food security data,
@@ -106,45 +105,29 @@ class DataBridgesShapes:
         ...     'WFP_API_CLIENT_SECRET': 'your-client-secret',
         ...     'DATABRIDGES_API_KEY': 'optional-databridges-key'
         ... }
-        >>> client = DataBridgesShapes(config)
+        >>> client = DataBridgesKnots(config)
 
         >>> # Initialize from environment variables
         >>> from data_bridges_knots.client import config_from_env
-        >>> client = DataBridgesShapes(config_from_env())
+        >>> client = DataBridgesKnots(config_from_env())
     """
 
     def __init__(self, yaml_config_path, env="prod", api_version="v1"):
-
-        warnings.warn(
-            (
-                "\n[FUTURE WARNING]\n"
-                "DataBridgesShapes will be renamed to 'DataBridgesKnots' "
-                "in version 4.0.0 (next major release, scheduled for 1 July 2026).\n"
-                "Please update your imports accordingly.\n"
-            ),
-            FutureWarning,
-            stacklevel=2,
-        )
-
-        # Initialize instance variables
         self.api_version = api_version
         self.env = env
         self.xlsform = None
 
-        # Load and validate config once
         self.config = self._load_config(yaml_config_path)
         self._validate_config(self.config)
-
-        # Setup authentication and extract API key
         self.configuration = self._setup_configuration_and_authentication(self.config)
         self.data_bridges_api_key = self.config.get("DATABRIDGES_API_KEY", "")
 
     def __repr__(self):
-        return f"DataBridgesShapes(host='{self.configuration.host}', env='{self.env}'), api_version='{self.api_version}'"
+        return f"DataBridgesKnots(host='{self.configuration.host}', env='{self.env}'), api_version='{self.api_version}'"
 
     def __str__(self):
         return (
-            f"DataBridgesShapes\n"
+            f"DataBridgesKnots\n"
             f"  API Host: {self.configuration.host}\n"
             f"  Environment: {self.env}\n"
             f"\n"
@@ -1473,6 +1456,105 @@ class DataBridgesShapes:
                 )
                 raise
 
+    def get_prices(
+        self,
+        country_iso3: str,
+        start_date: Optional[str] = None,
+        end_date: Optional[str] = None,
+        page_size: int = 1000,
+        market_id: int = 0,
+        commodity_id: int = 0,
+        currency_id: int = 0,
+        price_flag: str = "",
+        latest_value_only: bool = False,
+    ) -> pd.DataFrame:
+        """Fetches market price data for a given country within a specified date range.
+
+        Args:
+            country_iso3 (str): The ISO 3-letter country code
+            start_date (str, optional): Start date in ISO format (e.g., '2022-01-01').
+                If None, defaults to today's date.
+            end_date (str, optional): End date in ISO format (e.g., '2022-01-01').
+                If None, defaults to today's date.
+            page_size (int, optional): Number of items per page. Defaults to 1000.
+            market_id (int, optional): Unique ID of a Market. Defaults to 0.
+            commodity_id (int, optional): The exact ID of a Commodity. Defaults to 0.
+            currency_id (int, optional): The exact ID of a currency. Defaults to 0.
+            price_flag (str, optional): Type of price data: [actual|aggregate|estimated|forecasted]. Defaults to ''.
+            latest_value_only (bool, optional): Whether to return only latest values. Defaults to False.
+
+        Returns:
+            pd.DataFrame: DataFrame containing market price data
+
+        Examples:
+            >>> client = DataBridgesShapes("data_bridges_api_config.yaml")
+            >>> # Basic usage with dates
+            >>> df_prices = client.get_prices("KEN", start_date="2025-01-01", end_date="2025-12-31")
+            >>> # Using additional filters
+            >>> df_prices = client.get_prices(
+            ...     "KEN",
+            ...     start_date="2025-01-01",
+            ...     market_id=123,
+            ...     commodity_id=456,
+            ...     price_flag="actual"
+            ... )
+        """
+        if start_date:
+            # Format the date according to RFC 3339 standard
+            start_date = date.fromisoformat(start_date).strftime(
+                "%Y-%m-%dT%H:%M:%S+01:00"
+            )
+        else:
+            start_date = date.today().strftime("%Y-%m-%dT%H:%M:%S+01:00")
+
+        if end_date:
+            # Format the date according to RFC 3339 standard
+            end_date = date.fromisoformat(end_date).strftime("%Y-%m-%dT%H:%M:%S+01:00")
+        else:
+            end_date = date.today().strftime("%Y-%m-%dT%H:%M:%S+01:00")
+
+        responses = []
+        total_items = 20
+        max_item = 0
+        page = 0
+        while total_items > max_item:
+            page += 1
+            with data_bridges_client.ApiClient(
+                self._setup_configuration_and_authentication(self.config)
+            ) as api_client:
+                api_instance = data_bridges_client.MarketPricesApi(api_client)
+                env = self.env
+
+                try:
+                    api_prices = api_instance.market_prices_price_monthly_get(
+                        country_code=country_iso3,
+                        market_id=market_id,
+                        commodity_id=commodity_id,
+                        currency_id=currency_id,
+                        price_flag=price_flag,
+                        format="json",
+                        page=page,
+                        env=env,
+                        start_date=start_date,
+                        end_date=end_date,
+                        latest_value_only=latest_value_only,
+                    )
+                    responses.extend(item.to_dict() for item in api_prices.items)
+                    total_items = api_prices.total_items
+                    logger.info("Fetching page %s/n", page)
+                    max_item = page * page_size
+                    time.sleep(1)
+                except ApiException as e:
+                    logger.error(
+                        "Exception when calling Market price data->market_prices_price_monthly_get: %s\n",
+                        e,
+                    )
+                    raise
+
+        df = pd.DataFrame(responses)
+        df = df.replace({np.nan: None})
+        return df
+    
     def get_mfi_surveys_base_data(
         self, survey_id=None, page: Optional[int] = 1, page_size=20
     ):
@@ -1527,7 +1609,71 @@ class DataBridgesShapes:
         pass
 
 
-DataBridgesShapes.get_prices = get_prices
+
+class DataBridgesShapes(DataBridgesKnots):
+    """Interface to the Data Bridges API.
+
+    Provides methods for fetching market prices, exchange rates, food security data,
+    commodities, and more. Can be initialized from a YAML file, a dictionary, or
+    environment variables.
+
+    Args:
+        yaml_config_path (str | dict): Either:
+            - Path to a YAML configuration file (str), or
+            - Configuration dictionary (e.g. .env) with required keys: WFP_API_CLIENT_ID,
+              WFP_API_CLIENT_SECRET, and optionally DATABRIDGES_API_KEY
+        env (str, optional): Environment to use ('prod' or 'dev'). Defaults to "prod".
+        api_version (str, optional): Data Bridges API version to use. Defaults to "v1" (current version)
+
+
+    Examples:
+        >>> # Initialize with YAML file
+        >>> client = DataBridgesShapes("data_bridges_api_config.yaml")
+        >>> df_prices = client.get_prices("KEN", "2025-09-01")
+
+        >>> # Initialize with dictionary
+        >>> config = {
+        ...     'WFP_API_CLIENT_ID': 'your-client-id',
+        ...     'WFP_API_CLIENT_SECRET': 'your-client-secret',
+        ...     'DATABRIDGES_API_KEY': 'optional-databridges-key'
+        ... }
+        >>> client = DataBridgesShapes(config)
+
+        >>> # Initialize from environment variables
+        >>> from data_bridges_knots.client import config_from_env
+        >>> client = DataBridgesShapes(config_from_env())
+    """
+
+    def __init__(self, yaml_config_path, env="prod", api_version="v1"):
+
+        warnings.warn(
+            (
+                "\n[FUTURE WARNING]\n"
+                "DataBridgesShapes will be renamed to 'DataBridgesKnots' "
+                "in version 4.0.0 (next major release, scheduled for 1 July 2026).\n"
+                "Please update your imports accordingly.\n"
+            ),
+            FutureWarning,
+            stacklevel=2,
+        )
+
+        super().__init__(*args, **kwargs)
+
+    def __repr__(self):
+        return f"DataBridgesShapes(host='{self.configuration.host}', env='{self.env}'), api_version='{self.api_version}'"
+
+    def __str__(self):
+        return (
+            f"DataBridgesShapes\n"
+            f"  API Host: {self.configuration.host}\n"
+            f"  Environment: {self.env}\n"
+            f"\n"
+            f"Brought to you with <3 by WFP VAM"
+        )
+
+
+
+# DataBridgesShapes.get_prices = get_prices
 
 if __name__ == "__main__":
     pass
